centralize validation logic
