+++ bindings/python/zen.pyi
@@ -43,7 +43,6 @@ def compile_expression(expression: str) -> Expression: ...
 def compile_unary_expression(expression: str) -> Expression: ...
 
 
-
 class Expression:
     def evaluate(self, ctx: Optional[dict] = None) -> Any: ...
